Decompile more toys and move func_803485F0 to test_menu
- Decompile toy_80347C2C
- Decompile toy_8034819C
- Decompile toy_80348214
- Decompile toy_803483AC
- Decompile toy_80348418
- Move func_803485F0 toys -> test_menu

+++ src/app/code_7CF30.c
@@ -616,7 +616,7 @@ void func_802F7178(GyrocopterData* arg0, Mtx4F* arg1) {
 
 void func_802F7224(GyrocopterData* arg0) {
     GyrocopterData_Unk690* var_s0;
-    s32 sp100;
+    f32 sp100;
     f32 spFC;
     f32 spF8;
     s32 spF4;
