+++ src/app/toys.h
@@ -5,6 +5,6 @@
 
 void toyLoad(LevelTOYS*);
 void toy_803483AC(void);
-void toy_80348418(s32, s32, f32, f32, s32);
+void toy_80348418(s32, f32, f32, f32, s32);
 
 #endif // APP_TOYS_H
